Create AuthenticationBaseCache
