+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/cache/AuthenticationResultCache.java
@@ -20,14 +20,13 @@ package org.wso2.carbon.identity.application.authentication.framework.cache;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.wso2.carbon.identity.core.cache.BaseCache;
 import org.wso2.carbon.identity.core.util.IdentityUtil;
 
 /**
  * This cache keeps the information about the authentication result from the framework.
  */
 public class AuthenticationResultCache extends
-        BaseCache<AuthenticationResultCacheKey, AuthenticationResultCacheEntry> {
+        AuthenticationBaseCache<AuthenticationResultCacheKey, AuthenticationResultCacheEntry> {
 
     private static Log log = LogFactory.getLog(AuthenticationResultCache.class);
