+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/dao/LongWaitStatusDAO.java
@@ -31,5 +31,4 @@ public interface LongWaitStatusDAO {
     void removeWaitStatus(String waitKey) throws FrameworkException;
 
     LongWaitStatus getWaitStatus(String waitKey) throws FrameworkException;
-
 }
